SEO optimization, design improvements
Had rewritten all stuff which was not mine. Had changed the font from Roboto to Montserrat. Have changed the event section look. Rewrote the FAQ to be less of a keyword stuffer. Had deleted all the links in the /List, as it might be the reason for my website to be blocked on Bing (too many links with no content dragging people away).